# docs/api/clario-dr-service.openapi.yaml
openapi: 3.1.0
info:
  title: ClarioDR Service API
  version: 1.0.0
  summary: Disaster recovery control-plane and agent-ingest contract.
  description: |
    Phase 1 contract foundation for the existing ClarioDR API. The gateway
    exposes the tenant control plane at `/api/v1/dr` and gates it with the
    `suite.datastream` entitlement. Mutating control-plane writes stage DR
    lifecycle events through `backend/internal/events/outbox` before commit.
  x-contract-phase: phase-1-foundation
  x-gateway-contract:
    id: clario-dr-service
    version: 1.0.0
    api_version: v1
    fail_closed: false
    fail_closed_profiles:
      standard: false
      regulated: true
    regulated_posture:
      env:
        - DR_DEPLOYMENT_PROFILE=regulated
        - DR_REGULATED_MODE=true
      startup_validation: fail_closed
      readiness_check: dr_runtime_config
      required_runtime:
        - Vault Transit
        - MinIO WORM recovery-point bucket
        - Redis leader election
        - Kafka event bus
        - mTLS agent ingest
        - failover driver
        - integration credential encryption key
        - command recovery driver
  x-route-sources:
    - backend/cmd/clario-dr-service/main.go
    - backend/internal/dr/handler/handler.go
    - backend/internal/dr/enroll/handler.go
    - backend/internal/dr/ingest/handler.go
    - backend/cmd/clario-dr-service/resilience.go
    - backend/cmd/clario-dr-service/intelligence.go
    - backend/cmd/clario-dr-service/orchestration.go
...
servers:
...
security:
  - bearerAuth: []
    tenantContext: []
...
paths:
...
  /recovery-tiers:
    get:
      tags: [RecoveryTier]
      operationId: listDRRecoveryTiers
      summary: List built-in recovery objective tiers.
      x-required-permission: dr:read
      responses:
        "200":
          description: Built-in recovery tiers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryTierListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /recovery-tiers/coverage:
    get:
      tags: [RecoveryTier]
      operationId: getDRRecoveryTierCoverage
      summary: Summarize protected-site objective fit across the recovery-tier catalog.
      x-required-permission: dr:read
      responses:
        "200":
          description: Recovery-tier coverage summary.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /recovery-tiers/sites/{site}/recommend:
    get:
      tags: [RecoveryTier]
      operationId: recommendDRRecoveryTierForSite
      summary: Recommend a recovery tier for one protected site from configured objectives.
      x-required-permission: dr:read
      parameters:
        - name: site
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Site recovery-tier recommendation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/RecoveryTierSiteRecommendation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /recovery-tiers/{tier}:
    get:
      tags: [RecoveryTier]
      operationId: getDRRecoveryTier
      summary: Get one built-in recovery objective tier.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RecoveryTier"
      responses:
        "200":
          description: Recovery tier profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryTierProfileEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /recovery-tiers/recommend:
    post:
      tags: [RecoveryTier]
      operationId: recommendDRRecoveryTier
      summary: Recommend the lowest recovery tier that satisfies requested objectives.
      x-required-permission: dr:read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryTierRecommendationRequest"
      responses:
        "200":
          description: Recommended recovery tier.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryTierRecommendationEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /recovery-points/{id}/instant-recovery:
    post:
      tags: [InstantRecovery]
      operationId: startDRInstantRecovery
      summary: Start an instant recovery COW session over a sealed recovery point.
      x-required-permission: dr:failover
      x-outbox:
        topic: datastream.dr.events
        event_type: com.clario360.datastream.dr.instant.started
      parameters:
        - $ref: "#/components/parameters/InstantRecoveryPointID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartInstantRecoveryRequest"
      responses:
        "201":
          description: Instant recovery session started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantSessionEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/BadRequest"
  /instant-sessions/{id}:
    get:
      tags: [InstantRecovery]
      operationId: getDRInstantSession
      summary: Get an instant recovery session and hydration progress.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/InstantSessionID"
      responses:
        "200":
          description: Instant recovery progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantProgressEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /instant-sessions/{id}/chunks/{index}:
    get:
      tags: [InstantRecovery]
      operationId: readDRInstantChunk
      summary: Read raw bytes from an instant recovery COW chunk.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/InstantSessionID"
        - $ref: "#/components/parameters/ChunkIndex"
      responses:
        "200":
          description: Raw chunk bytes.
          headers:
            X-Clario-DR-Session-ID:
              schema:
                $ref: "#/components/schemas/UUID"
            X-Clario-DR-Chunk-Index:
              schema:
                type: integer
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [InstantRecovery]
      operationId: writeDRInstantChunk
      summary: Write raw bytes into an instant recovery COW overlay chunk.
      x-required-permission: dr:failover
      parameters:
        - $ref: "#/components/parameters/InstantSessionID"
        - $ref: "#/components/parameters/ChunkIndex"
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "204":
          description: Overlay chunk written.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /instant-sessions/{id}/finalize:
    post:
      tags: [InstantRecovery]
      operationId: finalizeDRInstantSession
      summary: Begin finalizing a ready instant recovery session into a standalone sealed copy.
      x-required-permission: dr:failover
      x-outbox:
        topic: datastream.dr.events
        event_type: com.clario360.datastream.dr.instant.finalized
      parameters:
        - $ref: "#/components/parameters/InstantSessionID"
      responses:
        "202":
          description: Instant recovery finalization accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantSessionEnvelope"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantContext:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Optional fallback when tenant context is not present in the JWT.
    agentMutualTLS:
      type: mutualTLS
  parameters:
...
    RecoveryTier:
      name: tier
      in: path
      required: true
      schema:
        type: string
        enum: [bronze, silver, gold, platinum]
    InstantRecoveryPointID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InstantSessionID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ChunkIndex:
      name: index
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication or tenant context missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated caller is not authorized.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Invalid state or already existing resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: Request was refused by a safety or semantic guard.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        details: true
        request_id:
          type: string
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    StartInstantRecoveryRequest:
      type: object
      additionalProperties: false
      properties:
        group_id:
          $ref: "#/components/schemas/UUID"
...
    RecoveryTierRecommendationRequest:
      type: object
      required: [rto_seconds, rpo_seconds]
      additionalProperties: false
      properties:
        rto_seconds:
          type: integer
          format: int64
          minimum: 1
        rpo_seconds:
          type: integer
          format: int64
          minimum: 1
        business_critical:
          type: boolean
        mission_critical:
          type: boolean
        regulated_data:
          type: boolean
        ransomware_sensitive:
          type: boolean
    RecoveryTierProfile:
      type: object
      required:
        - tier
        - rto_seconds
        - rpo_seconds
        - recommended_strategy
        - minimum_validation_cadence_seconds
        - requires_cyber_vault
        - requires_clean_room
        - capabilities
        - notes
      properties:
        tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        rto_seconds:
          type: integer
          format: int64
          minimum: 1
        rpo_seconds:
          type: integer
          format: int64
          minimum: 1
        recommended_strategy:
          type: string
          enum: [backup_restore, pilot_light, warm_standby, active_active]
        minimum_validation_cadence_seconds:
          type: integer
          format: int64
          minimum: 1
        requires_cyber_vault:
          type: boolean
        requires_clean_room:
          type: boolean
        capabilities:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            type: string
...
    RecoveryTierSiteRecommendation:
      type: object
      required:
        - tier
        - strategy
        - rto_objective_seconds
        - rpo_objective_seconds
        - validation_cadence_seconds
        - requires_cyber_vault
        - requires_clean_room
      properties:
        tier:
          type: string
          enum: [bronze, silver, gold, platinum]
        strategy:
          type: string
          enum: [backup_restore, pilot_light, warm_standby, active_active]
        rto_objective_seconds:
          type: integer
          minimum: 1
        rpo_objective_seconds:
          type: integer
          minimum: 1
        validation_cadence_seconds:
          type: integer
          minimum: 1
        requires_cyber_vault:
          type: boolean
        requires_clean_room:
          type: boolean
        capabilities:
          type: array
          items:
            type: string
        notes:
          type: array
          items:
            type: string
...
    InstantSession:
      type: object
      required: [id, tenant_id, recovery_point_id, state, chunks_total, chunks_hydrated, chunk_size, overlay_location, started_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        recovery_point_id:
          $ref: "#/components/schemas/UUID"
        group_id:
          $ref: "#/components/schemas/UUID"
        state:
          type: string
          enum: [HYDRATING, READY, FINALIZING, FINALIZED, FAILED]
        chunks_total:
          type: integer
          minimum: 0
        chunks_hydrated:
          type: integer
          minimum: 0
        chunk_size:
          type: integer
          minimum: 0
        overlay_location:
          type: string
        finalized_location:
          type: string
        last_error:
          type: string
        started_at:
          $ref: "#/components/schemas/Timestamp"
        ready_at:
          $ref: "#/components/schemas/Timestamp"
        finalized_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    InstantProgress:
      type: object
      required: [session, percent_complete]
      properties:
        session:
          $ref: "#/components/schemas/InstantSession"
        percent_complete:
          type: number
          minimum: 0
          maximum: 100
...
    RecoveryTierProfileEnvelope:
      $ref: "#/components/schemas/DataEnvelopeRecoveryTierProfile"
    RecoveryTierListEnvelope:
      $ref: "#/components/schemas/DataEnvelopeRecoveryTierList"
    RecoveryTierRecommendationEnvelope:
      $ref: "#/components/schemas/DataEnvelopeRecoveryTierRecommendation"
    InstantSessionEnvelope:
      $ref: "#/components/schemas/DataEnvelopeInstantSession"
    InstantProgressEnvelope:
      $ref: "#/components/schemas/DataEnvelopeInstantProgress"
...
    DataEnvelopeRecoveryTierProfile:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RecoveryTierProfile"
    DataEnvelopeRecoveryTierList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RecoveryTierProfile"
    DataEnvelopeRecoveryTierRecommendation:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/RecoveryTierProfile"
    DataEnvelopeInstantSession:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/InstantSession"
    DataEnvelopeInstantProgress:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/InstantProgress"
...